projet6_TALA_SOUZA_KOUNDJO/requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pdfminer.six>=20221105
python-docx==1.1.2
Werkzeug==3.0.4
httpx==0.27.0
python-dotenv==1.0.1
qdrant-client==1.7.3
tiktoken==0.6.0
spacy==3.7.2
fr-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.7.0/fr_core_news_lg-3.7.0-py3-none-any.whl
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl

# Chatbot RAG dependencies
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
faiss-cpu==1.9.0
sentence-transformers==3.3.1
